Extracted 'who', 'where', and 'levels' commands to files.
