+++ src/cmd/where.h
@@ -1,0 +1,14 @@
+//
+// Created by Sventovit on 08.09.2024.
+//
+
+#ifndef BYLINS_SRC_CMD_WHERE_H_
+#define BYLINS_SRC_CMD_WHERE_H_
+
+class CharData;
+
+void do_where(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_CMD_WHERE_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
